ai-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
youtube-transcript-api==0.6.2
openai==1.35.3
python-dotenv==1.0.1
pydantic==2.7.4
imageio-ffmpeg==0.5.1
